Solving a few errors

+++ Source/Camiao.cpp
@@ -9,7 +9,6 @@
  * @brief It contains the implementation of the class Camiao and its sub-classes
  */
 
-//Base class//
 Camiao::Camiao(const long long int& carg, const long long int& id) {
     CargaMax = carg;
     this->id = id;
@@ -26,46 +25,59 @@ void Camiao::removeTruck() {
     this-> CargaMax = -CargaMax;
 }
 
+
+//------------------------------------------------------------------------------------
 Congelado::Congelado(long long int carg, long long int Id) : Camiao(carg,Id){}
 
 Congelado::~Congelado() {}
 
 double Congelado::cal_preco() {
     //to implement after checking taxes format
-    long long int Carga_positive = CargaMax < 0?  -CargaMax: CargaMax;
+    long long int Carga_positive = CargaMax < 0 ?  -CargaMax : CargaMax;  // case the truck is not used anymore the max storage will be negative, but we want to calculate the profit with positive numbers
     return tax * (double)Carga_positive;
 }
 
-string Congelado::getType() const {return "Congelado";}
+string Congelado::getType() const {
+    return "Congelado";
+}
+
+//------------------------------------------------------------------------------------
 
 Perigoso::Perigoso(long long int carg, long long int Id) : Camiao(carg, Id){}
 
 Perigoso::~Perigoso() {}
 
 double Perigoso::cal_preco() {
-    return tax*(double)CargaMax;
+    return tax * (double)CargaMax;
 }
 
 string Perigoso::getType() const {return "Perigoso";}
+
+//------------------------------------------------------------------------------------
 
 Animals::Animals(long long int carg, long long int Id) : Camiao(carg, Id){}
 
 Animals::~Animals() {}
 
 double Animals::cal_preco() {
-    return tax*CargaMax;
+    return tax * (double)CargaMax;
 }
 
 string Animals::getType() const {
     return "Animal";}
+
+
+//------------------------------------------------------------------------------------
 
 string Base::getType() const {
     return "Base";
 }
 
 double Base::cal_preco() {
-    return tax*CargaMax;
+    return tax * (double)CargaMax;
 }
+
+//------------------------------------------------------------------------------------
 
 bool Compare(const Camiao* c1, const Camiao* c2){
     if (c1->getProfit() < c2->getProfit()) return true;
